requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.1
pyexecjs==1.5.1